Issues #5, Issues #8 Fixed detection "Normal" windows and Modal dialogs

+++ SmartContextMenu/Native/User32.cs
@@ -65,10 +65,10 @@ namespace SmartContextMenu.Native
         public static extern bool GetWindowPlacement(IntPtr hWnd, ref WindowPlacement lpwndpl);
 
         [DllImport("user32.dll")]
-        public static extern int SetWindowLong(IntPtr handle, int nIndex, int dwNewLong);
+        public static extern int SetWindowLong(IntPtr handle, int nIndex, long dwNewLong);
 
         [DllImport("user32.dll")]
-        public static extern int GetWindowLong(IntPtr handle, int nIndex);
+        public static extern long GetWindowLong(IntPtr handle, int nIndex);
 
         [DllImport("user32.dll")]
         public static extern int GetClassLong(IntPtr handle, int nIndex);
